Update utils.js - correct salvo_state_match code
Intended to fix issue 10.  When using a Platinum MX router, Companion updates Salvo state properly when initially querying router, when it sends XSALVO%ID#{4} - but it does not update salvo state when the router sends an update of the type XSALVO!ID#{4} .  (Notice the % vs the !)

It appears to me the original code at line 182 is looking for a salvo name of the type ID${Name} but not responding to a salvo number of the type ID#{Number} .  My router is sending Salvo updates with a number, not a name. I hope this change will capture both.  

I am a novice coder, please check for accuracy before accepting.